+++ examples/heijn/planner.py
@@ -1,6 +1,6 @@
 from mppiisaac.planner.mppi_isaac import MPPIisaacPlanner
 from mppiisaac.utils.config_store import ExampleConfig
-from mppiisaac.utils.mppi_utils import quaternion_to_yaw
+from mppiisaac.utils.conversions import quaternion_to_yaw
 import hydra
 import torch
 import zerorpc
